+++ Changes.md
@@ -182,6 +182,7 @@ Build
 Improvements
 ------------
 
+- ImageReader: promoted 'fileValid' from the internal OpenImageIOReader that reports per frame whether that files exists or not.
 - OpenImageIOReader: Added 'fileValid' Bool plug that returns a value per frame whether that files exists or not.
 
 1.2.9.0 (relative to 1.2.8.0)